# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAX SPEED RACING SHOP Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles
      responses:
        "200":
          description: Roles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    put:
      operationId: updateRole
      tags: [roles]
      summary: Update role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "200":
          description: Role updated
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
  /penjualan:
    get:
      operationId: listPenjualan
      tags: [penjualan]
      summary: List sales transactions
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: paymentMethod
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Penjualan"
    post:
      operationId: createPenjualan
      tags: [penjualan]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePenjualanRequest"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penjualan"
  /penjualan/{id}:
    get:
      operationId: getPenjualan
      tags: [penjualan]
      summary: Get a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penjualan"
    put:
      operationId: updatePenjualan
      tags: [penjualan]
      summary: Update a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePenjualanRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penjualan"
    delete:
      operationId: deletePenjualan
      tags: [penjualan]
      summary: Delete a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
        permissions:
          type: object
          additionalProperties: true
      required: [id, username, name, role]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
      required: [username, password, name, role]
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        permissions:
          type: object
          additionalProperties: true
      required: [id, name, permissions]
    CreateRoleRequest:
      type: object
      properties:
        name:
          type: string
        permissions:
          type: object
          additionalProperties: true
      required: [name, permissions]
...
    Penjualan:
      type: object
      properties:
        id:
          type: integer
        tanggal:
          type: string
        nomor:
          type: integer
        kodeTransaksi:
          type: string
        kodeBarang:
          type: string
        namaBarang:
          type: string
        brand:
          type: string
        harga:
          type: number
        qty:
          type: integer
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank, online_shop, kredit]
        nilaiCash:
          type: number
          nullable: true
        namaBank:
          type: string
          nullable: true
        nilaiBank:
          type: number
          nullable: true
        namaOnlineShop:
          type: string
          nullable: true
        nilaiOnlineShop:
          type: number
          nullable: true
        namaCustomer:
          type: string
          nullable: true
        nilaiKredit:
          type: number
          nullable: true
        statusCair:
          type: string
          enum: [pending, cair]
          nullable: true
        tanggalCair:
          type: string
          nullable: true
        hargaBeli:
          type: number
          nullable: true
        noFaktur:
          type: string
          nullable: true
        totalModal:
          type: number
          nullable: true
      required:
        - id
        - tanggal
        - nomor
        - kodeTransaksi
        - kodeBarang
        - namaBarang
        - brand
        - harga
        - qty
        - total
        - paymentMethod
    CreatePenjualanRequest:
      type: object
      properties:
        tanggal:
          type: string
        kodeBarang:
          type: string
        harga:
          type: number
        qty:
          type: integer
        paymentMethod:
          type: string
          enum: [cash, bank, online_shop, kredit]
        nilaiCash:
          type: number
          nullable: true
        namaBank:
          type: string
          nullable: true
        nilaiBank:
          type: number
          nullable: true
        namaOnlineShop:
          type: string
          nullable: true
        nilaiOnlineShop:
          type: number
          nullable: true
        namaCustomer:
          type: string
          nullable: true
        nilaiKredit:
          type: number
          nullable: true
        noFaktur:
          type: string
          nullable: true
      required: [tanggal, kodeBarang, harga, qty, paymentMethod]
...